+++ cypress/support/utils.tsx
@@ -31,5 +31,15 @@ export function cypressPassThroughTestsFactory(Component: ComponentType, props?:
     if (Component.displayName !== 'ObjectPageSection' && Component.displayName !== 'ObjectPageSubSection') {
       cy.findByTestId(testId).should('have.id', 'element-id');
     }
+
+export function mountWithCustomTagName<P extends { as?: keyof HTMLElementTagNameMap }>(
+  Component: ComponentType<P>,
+  props?: P
+) {
+  it('mount with custom tag name', () => {
+    const testId = 'component-to-be-tested';
+    const as = props?.as || 'header';
+    cy.mount(<Component as={as} data-testid={testId} {...props} />);
+    cy.get(`${as}[data-testid="${testId}"]`).should('be.visible');
   });
 }