+++ packages/main/src/components/ActionSheet/index.tsx
@@ -3,12 +3,14 @@
 import { isPhone } from '@ui5/webcomponents-base/dist/Device.js';
 import { useI18nBundle, useSyncRef } from '@ui5/webcomponents-react-base';
 import { clsx } from 'clsx';
-import React, { Children, forwardRef, ReactElement, useReducer, useRef } from 'react';
+import React, { forwardRef, ReactElement, useReducer, useRef } from 'react';
 import { createPortal } from 'react-dom';
 import { createUseStyles } from 'react-jss';
 import { ButtonDesign } from '../../enums';
 import { AVAILABLE_ACTIONS, CANCEL, X_OF_Y } from '../../i18n/i18n-defaults';
 import { addCustomCSSWithScoping } from '../../internal/addCustomCSSWithScoping';
+import { useCanRenderPortal } from '../../internal/ssr';
+import { flattenFragments } from '../../internal/utils';
 import { CustomThemingParameters } from '../../themes/CustomVariables';
 import { UI5WCSlotsNode } from '../../types';
 import {
@@ -141,7 +143,7 @@ const ActionSheet = forwardRef<ResponsivePopoverDomRef, ActionSheetPropTypes>((p
     modal,
     placementType,
     portalContainer,
-    showCancelButton,
+    showCancelButton = true,
     slot,
     style,
     verticalAlign,
@@ -159,9 +161,14 @@ const ActionSheet = forwardRef<ResponsivePopoverDomRef, ActionSheetPropTypes>((p
   const [focusedItem, setFocusedItem] = useReducer((_, action) => {
     return parseInt(action.target.dataset.actionBtnIndex);
   }, 0);
-  const childrenToRender = Children.toArray(children).filter(Boolean);
+  const childrenToRender = flattenFragments(children);
   const childrenArrayLength = childrenToRender.length;
   const childrenLength = isPhone() && showCancelButton ? childrenArrayLength + 1 : childrenArrayLength;
+
+  const canRenderPortal = useCanRenderPortal();
+  if (!canRenderPortal) {
+    return null;
+  }
 
   const handleCancelBtnClick = () => {
     popoverRef.current.close();
@@ -259,14 +266,9 @@ const ActionSheet = forwardRef<ResponsivePopoverDomRef, ActionSheetPropTypes>((p
         )}
       </div>
     </ResponsivePopover>,
-    portalContainer
+    portalContainer ?? document.body
   );
 });
-
-ActionSheet.defaultProps = {
-  showCancelButton: true,
-  portalContainer: document.body
-};
 
 ActionSheet.displayName = 'ActionSheet';