+++ packages/main/src/components/AnalyticalTable/ColumnHeader/index.tsx
@@ -1,9 +1,10 @@
 import { VirtualItem } from '@tanstack/react-virtual';
-import '@ui5/webcomponents-icons/dist/filter.js';
-import '@ui5/webcomponents-icons/dist/group-2.js';
-import '@ui5/webcomponents-icons/dist/sort-ascending.js';
-import '@ui5/webcomponents-icons/dist/sort-descending.js';
+import iconFilter from '@ui5/webcomponents-icons/dist/filter.js';
+import iconGroup from '@ui5/webcomponents-icons/dist/group-2.js';
+import iconSortAscending from '@ui5/webcomponents-icons/dist/sort-ascending.js';
+import iconSortDescending from '@ui5/webcomponents-icons/dist/sort-descending.js';
 import { ThemingParameters } from '@ui5/webcomponents-react-base';
+import { clsx } from 'clsx';
 import React, {
   CSSProperties,
   DragEventHandler,
@@ -11,6 +12,7 @@ import React, {
   KeyboardEventHandler,
   MouseEventHandler,
   ReactNode,
+  useRef,
   useState
 } from 'react';
 import { createUseStyles } from 'react-jss';
@@ -37,7 +39,6 @@ export interface ColumnHeaderProps {
   isRtl: boolean;
   children: ReactNode | ReactNode[];
   portalContainer: Element;
-  uniqueId: string;
   scaleXFactor?: number;
 
   //getHeaderProps()
@@ -80,6 +81,10 @@ const styles = {
     display: 'inline-block',
     position: 'absolute',
     color: ThemingParameters.sapContent_IconColor
+  },
+  selectAllCheckBoxContainer: {
+    display: 'flex',
+    justifyContent: 'center'
   }
 };
 
@@ -111,12 +116,12 @@ export const ColumnHeader: FC<ColumnHeaderProps> = (props: ColumnHeaderProps) =>
     onClick,
     onKeyDown,
     portalContainer,
-    uniqueId,
     scaleXFactor
   } = props;
 
   const isFiltered = column.filterValue && column.filterValue.length > 0;
   const [popoverOpen, setPopoverOpen] = useState(false);
+  const columnHeaderRef = useRef<HTMLDivElement>(null);
 
   const tooltip = (() => {
     if (headerTooltip) {
@@ -185,12 +190,10 @@ export const ColumnHeader: FC<ColumnHeaderProps> = (props: ColumnHeaderProps) =>
       setPopoverOpen(true);
     }
   };
-  const uniqueColumnId = `${column.id}-${uniqueId}`;
-
   if (!column) return null;
   return (
     <div
-      id={uniqueColumnId}
+      ref={columnHeaderRef}
       style={{
         position: 'absolute',
         top: 0,
@@ -238,7 +241,10 @@ export const ColumnHeader: FC<ColumnHeaderProps> = (props: ColumnHeaderProps) =>
             title={tooltip}
             wrapping={false}
             style={textStyle}
-            className={classes.text}
+            className={clsx(
+              classes.text,
+              id === '__ui5wcr__internal_selection_column' && classes.selectAllCheckBoxContainer
+            )}
             data-component-name={`AnalyticalTableHeaderHeaderContentContainer-${id}`}
           >
             {children}
@@ -248,9 +254,9 @@ export const ColumnHeader: FC<ColumnHeaderProps> = (props: ColumnHeaderProps) =>
             style={iconContainerDirectionStyles}
             data-component-name={`AnalyticalTableHeaderIconsContainer-${id}`}
           >
-            {isFiltered && <Icon name="filter" />}
-            {column.isSorted && <Icon name={column.isSortedDesc ? 'sort-descending' : 'sort-ascending'} />}
-            {column.isGrouped && <Icon name="group-2" />}
+            {isFiltered && <Icon name={iconFilter} />}
+            {column.isSorted && <Icon name={column.isSortedDesc ? iconSortDescending : iconSortAscending} />}
+            {column.isGrouped && <Icon name={iconGroup} />}
           </div>
         </div>
         {hasPopover && popoverOpen && (
@@ -259,7 +265,7 @@ export const ColumnHeader: FC<ColumnHeaderProps> = (props: ColumnHeaderProps) =>
             column={column}
             onSort={onSort}
             onGroupBy={onGroupBy}
-            uniqueColumnId={uniqueColumnId}
+            openerRef={columnHeaderRef}
             open={popoverOpen}
             setPopoverOpen={setPopoverOpen}
             portalContainer={portalContainer}