+++ packages/main/src/webComponents/TableCell/index.tsx
@@ -24,7 +24,14 @@ export interface TableCellPropTypes extends TableCellAttributes, CommonProps {
  *
  * <ui5-link href="https://sap.github.io/ui5-webcomponents/playground/components/Table" target="_blank">UI5 Web Components Playground</ui5-link>
  */
-const TableCell = withWebComponent<TableCellPropTypes, TableCellDomRef>('ui5-table-cell', [], [], [], []);
+const TableCell = withWebComponent<TableCellPropTypes, TableCellDomRef>(
+  'ui5-table-cell',
+  [],
+  [],
+  [],
+  [],
+  () => import('@ui5/webcomponents/dist/TableCell.js')
+);
 
 TableCell.displayName = 'TableCell';