+++ packages/main/src/webComponents/TimelineItem/index.tsx
@@ -59,7 +59,8 @@ const TimelineItem = withWebComponent<TimelineItemPropTypes, TimelineItemDomRef>
   ['icon', 'name', 'subtitleText', 'titleText'],
   ['nameClickable'],
   [],
-  ['name-click']
+  ['name-click'],
+  () => import('@ui5/webcomponents-fiori/dist/TimelineItem.js')
 );
 
 TimelineItem.displayName = 'TimelineItem';